scripts/requirements-backtest.txt
# Backtest scripts (VectorBT premium strategies, multi-backtest)
# Install: pip install -r scripts/requirements-backtest.txt
vectorbt>=0.26.0
pandas>=2.0
numpy>=1.24
requests>=2.28
python-dotenv>=1.0
# Optional: for yfinance fallback (NSE daily)
yfinance>=0.2.28
